requirements.txt
django
crispy-bootstrap5==0.6
django-crispy-forms==1.14.0
django-filter
django-htmx
django-tables2
faker==18.6.2
psycopg2-binary==2.9.6
gunicorn==20.0.4